Moved Violation to Refactoring. Clean test cases. Update CHANGELOG.md.
